requirements.txt
mne>=1.7
numpy>=2.0
scipy>=1.13
pandas>=2.2
matplotlib>=3.9
plotly>=5.23
streamlit>=1.37
pyyaml>=6.0
scikit-learn>=1.5
jinja2>=3.1
